sql/row: fix multi-range reads from PCR standby
Reads from tables with external row data (i.e. reads from a PCR standby
cluster) need to use the fixed timestamp specified by the external row
data. This timestamp might be different from the transaction timestamp,
so we were explicitly setting BatchRequest.Timestamp in
kv_batch_fetcher.

The KV API only allows BatchRequest.Timestamp to be set for
non-transactional requests (i.e. requests sent with a
NonTransactionalSender, which is a CrossRangeTxnWrapperSender in this
case). We were using a NonTransactionalSender, but this had two
problems:

1. CrossRangeTxnWrapperSender in turn sends the BatchRequest with a
   transactional sender, which again does not allow
   BatchRequest.Timestamp to be set.
2. CrossRangeTxnWrapperSender uses `kv.(*Txn).CommitInBatch`, which does
   not provide the 1-to-1 request-response guarantee required by
   txnKVFetcher. It is `kv.(*Txn).Send` which provides this guarantee.

Because of these two problems, whenever the txnKVFetcher would send a
multi-range-spanning BatchRequest to CrossRangeTxnWrapperSender, it
would either fail with a "transactional request must not set batch
timestamp" error or would return an unexpected number of responses,
violating the txnKVFetcher's assumed mapping from request to response.

To fix both these problems, instead of using a NonTransactionalSender,
change the txnKVFetcher to open a new root transaction with the correct
fixed timestamp, and then use txn.Send.

Fixes: #132608

Release note: None